Reappropriated class kwargs in svg code to be used for value of the elements. the class_ kwargs as such is no longer defined.
class is no longer defined for the elements because the class kwarg in the svg code is now used to save the value of each element for easy parsing on the website. If the class was to become relevant in the future, it can be indirectly deduced by scanning the id, because R1 e.g. is always a Resistor and L1 an inductor etc. To store the value of an element, one can still set it via the value_ keyword. The only change is that now before the svg code is created the string that is saved in the value_ argument is assigend to the class keyword in the svg Code