Created json files for UI_calculations for Circuit_resistor_task1
Name value pairs are similar to the normal json files for the resistance calculations. However due to the fact the R,I and U need to be specified - arrays of these values are used. Each name has normally three values: one for R one for U and one for I. Also relation type, result and equation are implemented.