Changed the name of the "paragraph" to "paragraph_Z"
This is in order to allow for differentiation between the paragraph that should be displayed while simplifying the circuit (paragraph_Z) and the paragraph that should be displayed while calculating the voltage and current (paragraph_UI). Paragraph_UI is not present at this moment but will be implemented as soon as possible.